url-recon/backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
# ── Database ──────────────────────────────────────────────────────────────────
# sqlalchemy      — ORM + query builder; talks to Postgres using Python classes
#                   instead of raw SQL strings. One URL swap switches databases.
# asyncpg         — The async Postgres wire driver. Fastest Python-to-Postgres
#                   driver available. Required by SQLAlchemy's async engine.
# greenlet        — Low-level coroutine library SQLAlchemy async needs internally.
#                   You never call it directly — it runs behind the scenes.
# python-dotenv   — Reads .env files and loads them into os.environ so
#                   credentials never have to be hardcoded in source files.
sqlalchemy==2.0.41
asyncpg==0.30.0
greenlet==3.2.1
python-dotenv==1.1.0
# ─────────────────────────────────────────────────────────────────────────────
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
cssselect2==0.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-users==15.0.5
fastapi-users-db-sqlalchemy==7.0.0
fonttools==4.62.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
makefun==1.16.0
MarkupSafe==3.0.3
pillow==12.1.1
pwdlib==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
python-whois==0.9.6
setuptools==78.1.1
six==1.17.0
starlette==0.52.1
tinycss2==1.5.1
tinyhtml5==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
weasyprint==68.1
webencodings==0.5.1
wheel==0.45.1
zopfli==0.4.1
